# mocks: Speed up generation of mock interfaces