--- a/cmd/frontend/internal/auth/userpasswd/gen.go
+++ b/cmd/frontend/internal/auth/userpasswd/gen.go
@@ -1,3 +1,0 @@
-package userpasswd
-
-//go:generate ../../../../../dev/mockgen.sh github.com/sourcegraph/sourcegraph/cmd/frontend/internal/auth/userpasswd -o mocks.go -i LockoutStore