--- a/cmd/symbols/fetcher/gen.go
+++ b/cmd/symbols/fetcher/gen.go
@@ -1,3 +1,0 @@
-package fetcher
-
-//go:generate ../../../dev/mockgen.sh github.com/sourcegraph/sourcegraph/cmd/symbols/gitserver -i GitServerClient -o mock_iface_test.go