--- a/cmd/symbols/internal/api/gen.go
+++ b/cmd/symbols/internal/api/gen.go
@@ -1,3 +1,0 @@
-package api
-
-//go:generate ../../../../dev/mockgen.sh github.com/sourcegraph/sourcegraph/cmd/symbols/gitserver -i GitServerClient -o mock_iface_test.go