--- a/dev/ci/go-build.sh
+++ b/dev/ci/go-build.sh
@@ -7,8 +7,6 @@ echo "--- go mod download"
 go mod download
 
 echo "--- go generate"
-export GOBIN="${PWD}/.bin"
-go install golang.org/x/tools/cmd/goimports
 go generate ./...
 
 echo "--- go install"