--- a/dev/sg/internal/generate/golang/golang.go
+++ b/dev/sg/internal/generate/golang/golang.go
@@ -93,9 +93,7 @@ func Generate(ctx context.Context, args []string, progressBar bool, verbosity Ou
 	// Determine which goimports we can use
 	var goimportsBinary string
 	if _, err := exec.LookPath("goimports"); err != nil {
-		// Install a local version of goimports - we do this whether we have a
-		// version of goimports or not because we need to feed it into the go-mockgen
-		// configuration file (which we don't yet template).
+		// Install goimports if not present
 		err = run.Cmd(ctx, "go", "install", "golang.org/x/tools/cmd/goimports").
 			Environ(os.Environ()).
 			Env(map[string]string{