--- a/docker-images/prometheus/cmd/prom-wrapper/mocks/gen.go
+++ b/docker-images/prometheus/cmd/prom-wrapper/mocks/gen.go
@@ -1,3 +1,0 @@
-package mocks
-
-//go:generate ../../../../../dev/mockgen.sh github.com/prometheus/client_golang/api/prometheus/v1 -i API -o prometheus_mock.go