--- a/enterprise/cmd/frontend/internal/app/gen.go
+++ b/enterprise/cmd/frontend/internal/app/gen.go
@@ -1,3 +1,0 @@
-package app
-
-//go:generate ../../../../../dev/mockgen.sh github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/app -i githubClient -o mock_github_client_test.go