--- a/enterprise/cmd/frontend/internal/codeintel/httpapi/gen.go
+++ b/enterprise/cmd/frontend/internal/codeintel/httpapi/gen.go
@@ -1,3 +1,0 @@
-package httpapi
-
-//go:generate ../../../../../../dev/mockgen.sh github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/codeintel/httpapi -i DBStore -i GitHubClient -o mock_iface_test.go