--- a/enterprise/cmd/frontend/internal/executorqueue/handler/gen.go
+++ b/enterprise/cmd/frontend/internal/executorqueue/handler/gen.go
@@ -1,3 +1,0 @@
-package handler
-
-//go:generate ../../../../../../dev/mockgen.sh github.com/sourcegraph/sourcegraph/internal/services/executors/store -i Store -o mock_iface_test.go