--- a/enterprise/cmd/frontend/internal/executorqueue/queues/batches/gen.go
+++ b/enterprise/cmd/frontend/internal/executorqueue/queues/batches/gen.go
@@ -1,3 +1,0 @@
-package batches
-
-//go:generate ../../../../../../../dev/mockgen.sh github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/executorqueue/queues/batches -i BatchesStore -o mock_iface_test.go