--- a/enterprise/cmd/frontend/internal/registry/stores/dbmocks/gen.go
+++ b/enterprise/cmd/frontend/internal/registry/stores/dbmocks/gen.go
@@ -1,3 +1,0 @@
-package dbmocks
-
-//go:generate ../../../../../../../dev/mockgen.sh github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/registry/stores -i ReleaseStore -i ExtensionStore -o mock_stores.go