--- a/enterprise/cmd/worker/internal/codeintel/janitor/gen.go
+++ b/enterprise/cmd/worker/internal/codeintel/janitor/gen.go
@@ -1,3 +1,0 @@
-package janitor
-
-//go:generate ../../../../../../dev/mockgen.sh github.com/sourcegraph/sourcegraph/enterprise/cmd/worker/internal/codeintel/janitor -i DBStore -i LSIFStore -o mock_iface_test.go