--- a/enterprise/internal/batches/reconciler/gen.go
+++ b/enterprise/internal/batches/reconciler/gen.go
@@ -1,3 +1,0 @@
-package reconciler
-
-//go:generate ../../../../dev/mockgen.sh github.com/sourcegraph/sourcegraph/enterprise/internal/batches/sources -i ChangesetSource -i ForkableChangesetSource -o mock_changeset_source_test.go