--- a/enterprise/internal/batches/sources/gen.go
+++ b/enterprise/internal/batches/sources/gen.go
@@ -1,4 +1,0 @@
-package sources
-
-//go:generate ../../../../dev/mockgen.sh github.com/sourcegraph/sourcegraph/enterprise/internal/batches/sources -i ChangesetSource -i SourcerStore -o mock_iface_test.go
-//go:generate ../../../../dev/mockgen.sh github.com/sourcegraph/sourcegraph/internal/extsvc/bitbucketcloud -i Client --prefix BitbucketCloud -o mock_bitbucketcloud_client_test.go