--- a/enterprise/internal/batches/syncer/gen.go
+++ b/enterprise/internal/batches/syncer/gen.go
@@ -1,3 +1,0 @@
-package syncer
-
-//go:generate ../../../../dev/mockgen.sh github.com/sourcegraph/sourcegraph/enterprise/internal/batches/syncer -i SyncStore -o mock_iface_test.go