--- a/enterprise/internal/database/mock_gen.go
+++ b/enterprise/internal/database/mock_gen.go
@@ -1,3 +1,0 @@
-package database
-
-//go:generate ../../../dev/mockgen.sh github.com/sourcegraph/sourcegraph/enterprise/internal/database -o mocks.go -i EnterpriseDB -i CodeMonitorStore -i PermsStore