--- a/enterprise/internal/insights/background/gen.go
+++ b/enterprise/internal/insights/background/gen.go
@@ -1,3 +1,0 @@
-package background
-
-//go:generate ../../../../dev/mockgen.sh github.com/sourcegraph/sourcegraph/enterprise/internal/insights/background -i RepoStore -o mock_repo_store.go