--- a/enterprise/internal/insights/compression/gen.go
+++ b/enterprise/internal/insights/compression/gen.go
@@ -1,3 +1,0 @@
-package compression
-
-//go:generate ../../../../dev/mockgen.sh github.com/sourcegraph/sourcegraph/enterprise/internal/insights/compression -i CommitStore -o mock_commit_store.go