--- a/gen.go
+++ b/gen.go
@@ -1,0 +1,3 @@
+package sourcegraph
+
+//go:generate ./dev/mockgen.sh