--- a/gen.go
+++ b/gen.go
@@ -1,4 +1,4 @@
 package sourcegraph
 
 // Keep this version in sync with go.mod
-//go:generate go run github.com/derision-test/go-mockgen/cmd/go-mockgen@v1.3.0
+//go:generate go run github.com/derision-test/go-mockgen/cmd/go-mockgen@v1.3.1