--- a/internal/codeintel/autoindexing/background/scheduler/gen.go
+++ b/internal/codeintel/autoindexing/background/scheduler/gen.go
@@ -1,3 +1,0 @@
-package scheduler
-
-//go:generate ../../../../../dev/mockgen.sh github.com/sourcegraph/sourcegraph/internal/codeintel/autoindexing/background/scheduler -i DBStore -i PolicyMatcher -o mock_iface_test.go