--- a/internal/codeintel/autoindexing/gen.go
+++ b/internal/codeintel/autoindexing/gen.go
@@ -1,3 +1,0 @@
-package autoindexing
-
-//go:generate ../../../dev/mockgen.sh github.com/sourcegraph/sourcegraph/internal/codeintel/autoindexing -i DBStore -i GitServerClient -i RepoUpdaterClient -i InferenceService -o mock_iface_test.go