--- a/internal/codeintel/documents/gen.go
+++ b/internal/codeintel/documents/gen.go
@@ -1,3 +1,0 @@
-package documents
-
-//go:generate ../../../dev/mockgen.sh github.com/sourcegraph/sourcegraph/internal/codeintel/documents/internal/store -i Store -o mock_iface_test.go