… cmd/frontend/internal/auth/userpasswd/mocks.go
… cmd/symbols/fetcher/mock_iface_test.go
… cmd/symbols/internal/api/mock_iface_test.go
… dev/team/mock.go
… docker-images/prometheus/cmd/prom-wrapper/mocks/prometheus_mock.go
… enterprise/cmd/executor/internal/command/mock_runner_test.go
… enterprise/cmd/executor/internal/command/mock_store_test.go
… enterprise/cmd/executor/internal/worker/mock_command_runner_test.go
… enterprise/cmd/executor/internal/worker/mock_store_test.go
… enterprise/cmd/frontend/internal/app/mock_github_client_test.go
… enterprise/cmd/frontend/internal/codeintel/httpapi/mock_iface_test.go
… enterprise/cmd/frontend/internal/codeintel/resolvers/mock_iface_test.go
… enterprise/cmd/frontend/internal/codeintel/resolvers/mock_position_adjuster_test.go
… enterprise/cmd/frontend/internal/codeintel/resolvers/mocks/mock_query.go
… enterprise/cmd/frontend/internal/codeintel/resolvers/mocks/mock_resolver.go
… enterprise/cmd/frontend/internal/executorqueue/handler/mock_iface_test.go
… enterprise/cmd/frontend/internal/executorqueue/queues/batches/mock_iface_test.go
… enterprise/cmd/frontend/internal/registry/stores/dbmocks/mock_stores.go
… enterprise/cmd/precise-code-intel-worker/internal/worker/mock_iface_test.go
… enterprise/cmd/precise-code-intel-worker/internal/worker/mock_worker_store_test.go
… enterprise/cmd/worker/internal/codeintel/indexing/mock_iface_test.go
… enterprise/cmd/worker/internal/codeintel/indexing/mock_scanner_test.go
… enterprise/cmd/worker/internal/codeintel/indexing/mock_worker_store_test.go
… enterprise/cmd/worker/internal/codeintel/janitor/mock_iface_test.go
… enterprise/internal/batches/reconciler/mock_changeset_source_test.go
… enterprise/internal/batches/sources/mock_bitbucketcloud_client_test.go
… enterprise/internal/batches/sources/mock_iface_test.go
… enterprise/internal/batches/syncer/mock_iface_test.go
… enterprise/internal/database/mocks.go
… enterprise/internal/insights/background/mock_repo_store.go