@efritz 2022-06-06
@bobheadxi PTAL, changed some error handling in sg migrate that seemed very wrong.