@efritz 2022-06-07
In CI: the `go generate` step on `Go build` took 3m4s on `main`, compared to 1m4s on this branch.